refactor: 데이터 파일 분리
