+++ src/temp/responseData.json
@@ -1,22 +1,3 @@
 {
-    "recommendation": "선택하신 지역의 오늘 날씨는 맑음이며, 이후 4일간의 날씨는 맑음/흐림/비/흐림이 될 예정입니다. 캐주얼한 스타일을 원하셨으니 옷장의 데이터를 기반으로, 브이넥 가디건과 ㄹ",
-    "weather": {
-    "cod": "200",
-    "message": 0,
-    "cnt": 40,
-    "city": {
-    "id": 1835848,
-    "name": "Seoul",
-    "coord": {
-    "lat": 37.5665,
-    "lon": 126.978
-    },
-    "country": "KR",
-    "population": 10349312,
-    "timezone": 32400,
-    "sunrise": 1740607613,
-    "sunset": 1740648191
-    },
-    "condition": "Clear"
-    }
-    }
+    "recommendation": "선택하신 지역의 오늘 날씨는 맑음이며, 이후 4일간의 날씨는 맑음/흐림/비/흐림이 될 예정입니다. 캐주얼한 스타일을 원하셨으니 옷장의 데이터를 기반으로, 브이넥 가디건과 ㄹ"
+}
